requirements.txt
fastapi
requests
jsonlines
python-dotenv
uvicorn
pandas
tqdm
sqlite-vec
sentence-transformers
spacy
openai
argon2-cffi